Chapter 5 Use Link
Use nextjs Link component instead of a tag to prevent full page refresh.
nextjs codesplits so that if there is an error on a page other parts of
the application will still work.
React SPA loads all app code on initial load.

Where Link appears, nextjs prefetches the code for the linked route
in the background.